# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  BasicResource:
    description: |-
      These mutex/semaphores can represent external resources that job requires and can be used to determine
      concurrency of jobs. For example, a job may need a license key to connect to a third party service, and
      it may only accept upto five connections that can be allocated via resources.
    properties:
      category:
        description: Category can be used to represent grouping of resources
        type: string
        x-go-name: Category
      description:
        description: Description of resource
        type: string
        x-go-name: Description
      extract_config:
        $ref: '#/definitions/ResourceCriteriaConfig'
      platform:
        description: Platform can be OS platform or target runtime
        type: string
        x-go-name: Platform
      resource_type:
        description: ResourceType defines type of resource such as Device, CPU, Memory
        type: string
        x-go-name: ResourceType
      tags:
        description: Tags can be used as tags for resource matching
        items:
          type: string
        type: array
        x-go-name: Tags
      value:
        description: Value consumed, e.g. it will be 1 for mutex, semaphore but can
          be higher number for other quota system
        format: int64
        type: integer
        x-go-name: Value
    title: BasicResource defines common properties of a resource, which can be used
      to implement mutex/semaphores for a job.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
...
  JobDefinition:
    description: |-
      The workflow progresses based on the exit codes of tasks, determining the subsequent task to execute.
      Each task definition encapsulates a job's specifics, and upon receiving a new job request, an instance of
      this job is initiated through JobExecution.
    properties:
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      cron_trigger:
        description: CronTrigger can be used to run the job periodically
        type: string
        x-go-name: CronTrigger
      delay_between_retries:
        $ref: '#/definitions/Duration'
      description:
        description: Description of job
        type: string
        x-go-name: Description
      disabled:
        description: disabled is used to stop further processing of job, and it can
          be used during maintenance, upgrade or debugging.
        type: boolean
        x-go-name: Disabled
      hard_reset_after_retries:
        description: HardResetAfterRetries defines retry config when job is rerun
          and as opposed to re-running only failed tasks, all tasks are executed.
        format: int64
        type: integer
        x-go-name: HardResetAfterRetries
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_type:
        description: JobType defines a unique type of job
        type: string
        x-go-name: JobType
      job_variables:
        type: object
        x-go-name: NameValueVariables
      max_concurrency:
        description: MaxConcurrency defines max number of jobs that can be run concurrently
        format: int64
        type: integer
        x-go-name: MaxConcurrency
      methods:
        description: Methods is aggregation of task methods
        type: string
        x-go-name: Methods
      notify:
        additionalProperties:
          $ref: '#/definitions/JobNotifyConfig'
        type: object
        x-go-name: Notify
      organization_id:
        description: OrganizationID defines org who submitted the job
        type: string
        x-go-name: OrganizationID
      platform:
        description: Platform can be OS platform or target runtime and a job can be
          targeted for specific platform that can be used for filtering
        type: string
        x-go-name: Platform
      public_plugin:
        description: PublicPlugin means job is public plugin
        type: boolean
        x-go-name: PublicPlugin
      required_params:
        description: RequiredParams from job request (and plugin)
        items:
          type: string
        type: array
        x-go-name: RequiredParams
      resources:
        $ref: '#/definitions/BasicResource'
      retry:
        description: Retry defines max number of tries a job can be retried where
          it re-runs failed job
        format: int64
        type: integer
        x-go-name: Retry
      sem_version:
        description: SemVersion - semantic version is used for external version, which
          can be used for public plugins.
        type: string
        x-go-name: SemVersion
      tags:
        description: |-
          Tags are used to use specific followers that support the tags defined by ants.
          Tags is aggregation of task tags
        type: string
        x-go-name: Tags
      tasks:
        description: |-
          Tasks defines one to many relationships between job and tasks, where a job defines
          a directed acyclic graph of tasks that are executed for the job.
        items:
          $ref: '#/definitions/TaskDefinition'
        type: array
        x-go-name: Tasks
      timeout:
        $ref: '#/definitions/Duration'
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      url:
        description: URL defines url for job
        type: string
        x-go-name: URL
      user_id:
        description: UserID defines user who updated the job
        type: string
        x-go-name: UserID
    title: JobDefinition outlines a set of tasks arranged in a Directed Acyclic Graph
      (DAG), executed by worker entities.
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  JobNotifyConfig:
    description: JobNotifyConfig structure for notification config
    properties:
      recipients:
        items:
          type: string
        type: array
        x-go-name: Recipients
      when:
        $ref: '#/definitions/NotifyWhen'
    type: object
    x-go-package: plexobject.com/formicary/internal/types
...
  NotifyWhen:
    description: NotifyWhen type alias for when notify should be used
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
  ResourceCriteriaConfig:
    title: ResourceCriteriaConfig defines properties to extract from a resource into
      job variables.
...
  TaskDefinition:
    description: |-
      upon a new job request, a TaskExecution instance is initiated to carry out the task. The task details,
      including its method and tags, guide the dispatch of task requests to a compatible remote worker.
      Upon task completion, the outcomes are recorded in the database for reference.
    properties:
      after_script:
        description: AfterScript defines list of commands that are executed after
          main script for cleanup
        items:
          type: string
        type: array
        x-go-name: AfterScript
      allow_failure:
        description: AllowFailure means the task is optional and can fail without
          failing entire job
        type: boolean
        x-go-name: AllowFailure
      allow_start_if_completed:
        description: AllowStartIfCompleted  means the task is always run on retry
          even if it was completed successfully
        type: boolean
        x-go-name: AllowStartIfCompleted
      always_run:
        description: |-
          AlwaysRun means the task is always run on execution even if the job fails. For example, a required task fails (without
          AllowFailure), the job is aborted and remaining tasks are skipped but a task defined as `AlwaysRun` is run even if the job fails.
        type: boolean
        x-go-name: AlwaysRun
      artifact_ids:
        description: ArtifactIDs defines id of artifacts that are automatically downloaded
          for job-execution
        items:
          type: string
        type: array
        x-go-name: ArtifactIDs
      await_forked_tasks:
        description: AwaitForkedTasks defines list of jobs to wait for completion
        items:
          type: string
        type: array
        x-go-name: AwaitForkedTasks
      before_script:
        description: BeforeScript defines list of commands that are executed before
          main script
        items:
          type: string
        type: array
        x-go-name: BeforeScript
      created_at:
        description: CreatedAt job creation time
        format: date-time
        type: string
        x-go-name: CreatedAt
      delay_between_retries:
        $ref: '#/definitions/Duration'
      dependencies:
        description: Dependencies defines dependent tasks for downloading artifacts
        items:
          type: string
        type: array
        x-go-name: Dependencies
      description:
        description: Description of task
        type: string
        x-go-name: Description
      except:
        description: Except is used to shouldSkip task execution based on certain
          condition
        type: string
        x-go-name: Except
      fork_job_type:
        description: ForkJobType defines type of job to work
        type: string
        x-go-name: ForkJobType
      headers:
        additionalProperties:
          type: string
        description: Header defines HTTP headers
        type: object
        x-go-name: Headers
      host_network:
        description: HostNetwork defines kubernetes/docker config for host_network
        type: string
        x-go-name: HostNetwork
      id:
        description: |-
          gorm.Model
          ID defines UUID for primary key
        type: string
        x-go-name: ID
      job_definition_id:
        description: JobDefinitionID defines foreign key for JobDefinition
        type: string
        x-go-name: JobDefinitionID
      job_version:
        description: JobVersion defines job version
        type: string
        x-go-name: JobVersion
      messaging_reply_queue:
        type: string
        x-go-name: MessagingReplyQueue
      messaging_request_queue:
        type: string
        x-go-name: MessagingRequestQueue
      method:
        $ref: '#/definitions/TaskMethod'
      on_completed:
        description: OnCompleted defines next task to run based on completion
        type: string
        x-go-name: OnCompleted
      on_exit_code:
        additionalProperties:
          type: string
        description: OnExitCode defines next task to run based on exit code
        type: object
        x-go-name: OnExitCode
      on_failed:
        description: OnFailed defines next task to run based on failure
        type: string
        x-go-name: OnFailed
      report_stdout:
        description: ReportStdout is used to send stdout as a report
        type: boolean
        x-go-name: ReportStdout
      resources:
        $ref: '#/definitions/BasicResource'
      retry:
        description: Retry defines max number of tries a task can be retried where
          it re-runs failed tasks
        format: int64
        type: integer
        x-go-name: Retry
      script:
        description: Script defines list of commands to execute in container
        items:
          type: string
        type: array
        x-go-name: Script
      tags:
        description: |-
          Tags are used to use specific followers that support the tags defined by ants.
          For example, you may start a follower that processes payments and the task will be routed to that follower
        items:
          type: string
        type: array
        x-go-name: Tags
      task_type:
        description: TaskType defines type of task
        type: string
        x-go-name: TaskType
      timeout:
        $ref: '#/definitions/Duration'
      updated_at:
        description: UpdatedAt job update time
        format: date-time
        type: string
        x-go-name: UpdatedAt
      url:
        description: URL to use
        type: string
        x-go-name: URL
      variables:
        description: Transient properties -- these are populated when AfterLoad or
          Validate is called
        type: object
        x-go-name: NameValueVariables
      webhook:
        $ref: '#/definitions/Webhook'
    title: TaskDefinition outlines the work performed by worker entities. It specifies
      the task's parameters and,
    type: object
    x-go-package: plexobject.com/formicary/queen/types
...
  TaskMethod:
    description: |-
      The ant followers registers with the methods that they support and the task is then routed
      based on method, tags and concurrency of the ant follower.
    title: TaskMethod defines enum for method of task when communicating with remote
      ant follower.
    type: string
    x-go-package: plexobject.com/formicary/internal/types
...
host: https://formicary.io
info:
  contact:
    email: support@formicary.io
    name: Support
    url: https://formicary.io
  description: |-
    The formicary is a distributed orchestration engine based on `Leader-Follower`, `Pipes-Filter`, `Fork-Join` and `SEDA` design principles for
    executing a directed acyclic graph of tasks, which is also referred as a job workflow. A task represents a unit of work and a job definition is used to specify the task
    dependencies in the graph/workflow including configuration parameters and conditional logic.
  license:
    name: AGPL
    url: https://opensource.org/licenses/AGPL-3.0
  title: Formicary API.
  version: 0.0.1
paths:
...
  /api/jobs/definitions:
    get:
      operationId: queryJobDefinitions
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: JobType defines a unique type of job
        in: query
        name: job_type
        type: string
        x-go-name: JobType
      - description: Platform can be OS platform or target runtime and a job can be
          targeted for specific platform that can be used for filtering
        in: query
        name: platform
        type: string
        x-go-name: Platform
      - description: disabled is used to stop further processing of job, and it can
          be used during maintenance, upgrade or debugging.
        in: query
        name: disabled
        type: boolean
        x-go-name: Disabled
      - description: PublicPlugin means job is public plugin
        in: query
        name: public_plugin
        type: boolean
        x-go-name: PublicPlugin
      - description: Tags is aggregation of task tags, and it can be searched via
          `tags:in`
        in: query
        name: tags
        type: string
        x-go-name: Tags
      responses:
        "200":
          $ref: '#/responses/jobDefinitionQueryResponse'
      summary: Queries job definitions by criteria such as type, platform, etc.
      tags:
      - job-definitions
    post:
      operationId: postJobDefinition
      parameters:
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/JobDefinition'
      responses:
        "200":
          $ref: '#/responses/jobDefinition'
      summary: Uploads job definitions using JSON or YAML body based on content-type
        header.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}:
    delete:
      operationId: deleteJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Deletes the job-definition by id.
      tags:
      - job-definitions
    get:
      operationId: getJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/jobDefinition'
      summary: Finds the job-definition by id.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}/concurrency:
    put:
      operationId: updateConcurrencyJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      - format: int64
        in: formData
        name: concurrency
        type: integer
        x-go-name: Concurrency
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Updates the concurrency for job-definition by id to limit the maximum
        jobs that can be executed at the same time.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}/disable:
    post:
      operationId: disableJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: disables job-definition so that no new requests are executed while
        in-progress jobs are allowed to complete.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}/dot:
    get:
      operationId: dotJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/stringResponse'
      summary: Returns Graphviz DOT definition for the graph of tasks defined in the
        job.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}/dot.png:
    get:
      operationId: dotImageJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/byteResponse'
      summary: Returns Graphviz DOT image for the graph of tasks defined in the job.
      tags:
      - job-definitions
  /api/jobs/definitions/{id}/enable:
    post:
      operationId: enableJobDefinition
      parameters:
      - in: path
        name: id
        required: true
        type: string
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/emptyResponse'
      summary: Enables job-definition so that new requests can start processing.
      tags:
      - job-definitions
...
  /api/jobs/definitions/{type}/yaml:
    get:
      operationId: getYamlJobDefinition
      parameters:
      - in: path
        name: type
        required: true
        type: string
        x-go-name: Type
      responses:
        "200":
          $ref: '#/responses/jobDefinition'
      summary: Finds job-definition by type and returns response YAML format.
      tags:
      - job-definitions
  /api/jobs/plugins:
    get:
      operationId: queryPlugins
      parameters:
      - format: int64
        in: query
        name: page
        type: integer
        x-go-name: Page
      - format: int64
        in: query
        name: page_size
        type: integer
        x-go-name: PageSize
      - description: JobType defines a unique type of job
        in: query
        name: job_type
        type: string
        x-go-name: JobType
      - description: Platform can be OS platform or target runtime and a job can be
          targeted for specific platform that can be used for filtering
        in: query
        name: platform
        type: string
        x-go-name: Platform
      - description: disabled is used to stop further processing of job, and it can
          be used during maintenance, upgrade or debugging.
        in: query
        name: disabled
        type: boolean
        x-go-name: Disabled
      - description: PublicPlugin means job is public plugin
        in: query
        name: public_plugin
        type: boolean
        x-go-name: PublicPlugin
      - description: Tags is aggregation of task tags, and it can be searched via
          `tags:in`
        in: query
        name: tags
        type: string
        x-go-name: Tags
      responses:
        "200":
          $ref: '#/responses/jobDefinitionQueryResponse'
      summary: Queries job definitions by criteria such as type, platform, etc.
      tags:
      - job-definitions
...
produces:
- application/json
responses:
...
  byteResponse:
    description: Byte Array response body
    schema:
      items:
        format: uint8
        type: integer
      type: array
...
  emptyResponse:
    description: Empty response body
...
  jobDefinition:
    description: |-
      The job-definition defines DAG (directed acyclic graph) of tasks, which are executed by
      ant followers. The workflow of job uses task exit codes to define next task to execute.
    schema:
      $ref: '#/definitions/JobDefinition'
  jobDefinitionQueryResponse:
    description: Paginated results of jobDefinitions matching query
    schema:
      properties:
        Page:
          format: int64
          type: integer
        PageSize:
          format: int64
          type: integer
        Records:
          items:
            $ref: '#/definitions/JobDefinition'
          type: array
        TotalPages:
          format: int64
          type: integer
        TotalRecords:
          format: int64
          type: integer
      type: object
...
  stringResponse:
    description: String response body
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...